+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/CaptureJniLibrary.kt
@@ -395,6 +395,7 @@ internal object CaptureJniLibrary : IBridge, IStreamingReportProcessor {
         timestampMillis: Long,
         destinationPath: String,
         attributes: IClientAttributes,
+        runningState: String?,
     )
 
     /**