+++ platform/jvm/capture/src/main/kotlin/io/bitdrift/capture/events/lifecycle/AppExitLogger.kt
@@ -53,6 +53,7 @@ internal class AppExitLogger(
         private const val APP_EXIT_PSS_KEY = "_app_exit_pss"
         private const val APP_EXIT_RSS_KEY = "_app_exit_rss"
         private const val APP_EXIT_DESCRIPTION_KEY = "_app_exit_description"
+        private const val FOREGROUND_KEY = "foreground"
     }
 
     @SuppressLint("NewApi")
@@ -172,6 +173,7 @@ internal class AppExitLogger(
             APP_EXIT_PSS_KEY to this.pss.toString(),
             APP_EXIT_RSS_KEY to this.rss.toString(),
             APP_EXIT_DESCRIPTION_KEY to this.description.orEmpty(),
+            FOREGROUND_KEY to if (this.importance == RunningAppProcessInfo.IMPORTANCE_FOREGROUND) "1" else "0",
         )
     }